backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
